worldseo/requirements.txt
fastapi
uvicorn
feedparser
apscheduler
python-telegram-bot
requests
python-dotenv 